// packages/cactus-plugin-ledger-connector-corda/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface CordaNodeSshCredentials {
    /**
     * 
     * @type {string}
     * @memberof CordaNodeSshCredentials
     */
    'hostKeyEntry': string;
    /**
     * 
     * @type {string}
     * @memberof CordaNodeSshCredentials
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof CordaNodeSshCredentials
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof CordaNodeSshCredentials
     */
    'hostname': string;
    /**
     * 
     * @type {number}
     * @memberof CordaNodeSshCredentials
     */
    'port': number;
}
/**
 * 
 * @export
 * @interface CordaRpcCredentials
 */
export interface CordaRpcCredentials {
    /**
     * 
     * @type {string}
     * @memberof CordaRpcCredentials
     */
    'hostname': string;
    /**
     * 
     * @type {number}
     * @memberof CordaRpcCredentials
     */
    'port': number;
    /**
     * 
     * @type {string}
     * @memberof CordaRpcCredentials
     */
    'username': string;
    /**
     * 
     * @type {string}
     * @memberof CordaRpcCredentials
     */
    'password': string;
}
// ...
export interface CordappDeploymentConfig {
    /**
     * 
     * @type {CordaNodeSshCredentials}
     * @memberof CordappDeploymentConfig
     */
    'sshCredentials': CordaNodeSshCredentials;
    /**
     * 
     * @type {CordaRpcCredentials}
     * @memberof CordappDeploymentConfig
     */
    'rpcCredentials': CordaRpcCredentials;
    /**
     * The shell command to execute in order to start back up a Corda node after having placed new jars in the cordapp directory of said node.
     * @type {string}
     * @memberof CordappDeploymentConfig
     */
    'cordaNodeStartCmd': string;
    /**
     * The absolute file system path where the Corda Node is expecting deployed Cordapp jar files to be placed.
     * @type {string}
     * @memberof CordappDeploymentConfig
     */
    'cordappDir': string;
    /**
     * The absolute file system path where the corda.jar file of the node can be found. This is used to execute database schema migrations where applicable (H2 database in use in development environments).
     * @type {string}
     * @memberof CordappDeploymentConfig
     */
    'cordaJarPath': string;
    /**
     * The absolute file system path where the base directory of the Corda node can be found. This is used to pass in to corda.jar when being invoked for certain tasks such as executing database schema migrations for a deployed contract.
     * @type {string}
     * @memberof CordappDeploymentConfig
     */
    'nodeBaseDirPath': string;
}
// ...
export interface DeployContractJarsV1Request {
    /**
     * The list of deployment configurations pointing to the nodes where the provided cordapp jar files are to be deployed .
     * @type {Array<CordappDeploymentConfig>}
     * @memberof DeployContractJarsV1Request
     */
    'cordappDeploymentConfigs': Array<CordappDeploymentConfig>;
    /**
     * 
     * @type {Array<JarFile>}
     * @memberof DeployContractJarsV1Request
     */
    'jarFiles': Array<JarFile>;
}
// ...
export const FlowInvocationType = {
    TrackedFlowDynamic: 'TRACKED_FLOW_DYNAMIC',
    FlowDynamic: 'FLOW_DYNAMIC'
} as const;

export type FlowInvocationType = typeof FlowInvocationType[keyof typeof FlowInvocationType];
// ...
export interface InvokeContractV1Request {
    /**
     * The fully qualified name of the Corda flow to invoke
     * @type {string}
     * @memberof InvokeContractV1Request
     */
    'flowFullClassName': string;
    /**
     * 
     * @type {FlowInvocationType}
     * @memberof InvokeContractV1Request
     */
    'flowInvocationType': FlowInvocationType;
    /**
     * The list of arguments to pass in to the contract method being invoked.
     * @type {Array<JvmObject>}
     * @memberof InvokeContractV1Request
     */
    'params': Array<JvmObject>;
    /**
     * The amount of milliseconds to wait for a transaction receipt beforegiving up and crashing.
     * @type {number}
     * @memberof InvokeContractV1Request
     */
    'timeoutMs'?: number;
}
// ...
export interface JarFile {
    [key: string]: any;

    /**
     * 
     * @type {string}
     * @memberof JarFile
     */
    'filename': string;
    /**
     * Indicates whether the cordapp jar in question contains any embedded migrations that Cactus can/should execute between copying the jar into the cordapp directory and starting the node back up.
     * @type {boolean}
     * @memberof JarFile
     */
    'hasDbMigrations': boolean;
    /**
     * 
     * @type {string}
     * @memberof JarFile
     */
    'contentBase64': string;
}
// ...
export interface JvmObject {
    /**
     * 
     * @type {JvmTypeKind}
     * @memberof JvmObject
     */
    'jvmTypeKind': JvmTypeKind;
    /**
     * 
     * @type {any}
     * @memberof JvmObject
     */
    'primitiveValue'?: any;
    /**
     * 
     * @type {Array<JvmObject>}
     * @memberof JvmObject
     */
    'jvmCtorArgs'?: Array<JvmObject>;
    /**
     * 
     * @type {JvmType}
     * @memberof JvmObject
     */
    'jvmType': JvmType;
}


/**
 * Represents a reference to a JVM type (such as a Java class)
 * @export
 * @interface JvmType
 */
export interface JvmType {
    /**
     * 
     * @type {string}
     * @memberof JvmType
     */
    'fqClassName': string;
    /**
     * This parameter is used to specify that the function used to construct this JvmType is not a constructor function but instead is a factory function. Setting this parameter will cause the plugin to look up methods of the class denoted by fqClassName instead of its constructors.
     * @type {string}
     * @memberof JvmType
     */
    'constructorName'?: string;
    /**
     * 
     * @type {JvmObject}
     * @memberof JvmType
     */
    'invocationTarget'?: JvmObject;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const JvmTypeKind = {
    Primitive: 'PRIMITIVE',
    Reference: 'REFERENCE'
} as const;

export type JvmTypeKind = typeof JvmTypeKind[keyof typeof JvmTypeKind];